+++ test/LionWeb.Core.Test/Notification/Replicator/ReplicatorTests_Annotation.cs
@@ -140,6 +140,25 @@ public class ReplicatorTests_Annotation : ReplicatorTestsBase
         AssertEquals([originalPartition], [clonedPartition]);
     }
 
+    [TestMethod]
+    public void AnnotationDeleted_Multiple_not_matching_node_ids()
+    {
+        var deleted = new BillOfMaterials("deleted");
+        var nodeWithAnotherId = new BillOfMaterials("node-with-another-id");
+        var originalPartition = new Geometry("a");
+        originalPartition.AddAnnotations([new BillOfMaterials("bof"), deleted, nodeWithAnotherId]);
+
+        var clonedPartition = ClonePartition(originalPartition);
+
+        var notification = new AnnotationDeletedNotification(nodeWithAnotherId, originalPartition, 1,
+            new NumericNotificationId("annotationDeletedNotification", 0));
+
+        Assert.ThrowsExactly<InvalidNotificationException>(() =>
+        {
+            CreatePartitionReplicator(clonedPartition, notification);
+        });
+    }
+    
     /// <summary>
     /// This test confirms that no notification is generated from DetachFromParent method
     /// TODO: This is a known bug, we want to have a notification emitted.
@@ -158,7 +177,8 @@ public class ReplicatorTests_Annotation : ReplicatorTestsBase
 
         Assert.AreEqual(0, notificationObserver.Notifications.Count);
     }
-    
+
+
     #endregion
 
     #region AnnotationReplaced
@@ -276,6 +296,27 @@ public class ReplicatorTests_Annotation : ReplicatorTestsBase
 
         Assert.AreEqual(2, clonedPartition.GetAnnotations().Count);
         Assert.AreEqual(newAnnotation.GetId(), clonedPartition.GetAnnotations()[index].GetId());
+    }
+    
+    [TestMethod]
+    public void AnnotationReplaced_Multiple_not_matching_node_ids()
+    {
+        var replaced = new BillOfMaterials("replaced");
+        var nodeWithAnotherId = new BillOfMaterials("node-with-another-id");
+        var originalPartition = new Geometry("a");
+        originalPartition.AddAnnotations([new BillOfMaterials("bof"), replaced, nodeWithAnotherId]);
+
+        var clonedPartition = ClonePartition(originalPartition);
+
+        var index = 1;
+        var newAnnotation = new BillOfMaterials("new");
+        var notification = new AnnotationReplacedNotification(newAnnotation, nodeWithAnotherId, originalPartition,
+            index, new NumericNotificationId("annotationReplaced", 0));
+
+        Assert.ThrowsExactly<InvalidNotificationException>(() =>
+        {
+            CreatePartitionReplicator(clonedPartition, notification);
+        });
     }
 
     #endregion
@@ -405,6 +446,32 @@ public class ReplicatorTests_Annotation : ReplicatorTestsBase
        
         Assert.AreEqual(2, clonedPartition.Shapes[1].GetAnnotations().Count);
         Assert.AreEqual(moved.GetId(), clonedPartition.Shapes[1].GetAnnotations()[^1].GetId());
+    }
+
+    [TestMethod]
+    public void AnnotationMovedAndReplacedFromOtherParent_not_matching_node_ids()
+    {
+        var replaced = new BillOfMaterials("replaced");
+        var nodeWithAnotherId = new BillOfMaterials("node-with-another-id");
+        var line = new Line("line");
+        line.AddAnnotations([new BillOfMaterials("bof"), replaced, nodeWithAnotherId]);
+        
+        var origin = new CompositeShape("origin");
+        var moved = new BillOfMaterials("moved");
+        origin.AddAnnotations([moved]);
+        var originalPartition = new Geometry("a") { Shapes = [origin, line] };
+
+        var clonedPartition = ClonePartition(originalPartition);
+        
+        var newIndex = 1;
+        var oldIndex = 0;
+        var notification = new AnnotationMovedAndReplacedFromOtherParentNotification(line, newIndex, moved, origin, oldIndex, nodeWithAnotherId, 
+            new NumericNotificationId("AnnotationMovedAndReplacedFromOtherParent", 0));
+        
+        Assert.ThrowsExactly<InvalidNotificationException>(() =>
+        {
+            CreatePartitionReplicator(clonedPartition, notification);
+        });
     }
 
     #endregion 
@@ -606,6 +673,28 @@ public class ReplicatorTests_Annotation : ReplicatorTestsBase
 
         Assert.AreEqual(4, clonedPartition.GetAnnotations().Count);
         Assert.AreEqual(moved.GetId(), clonedPartition.GetAnnotations()[2].GetId());
+    }
+    
+    [TestMethod]
+    public void AnnotationMovedAndReplacedInSameParent_not_matching_node_ids()
+    {
+        var replaced = new BillOfMaterials("replaced");
+        var nodeWithAnotherId = new BillOfMaterials("node-with-another-id");
+        var moved = new BillOfMaterials("moved");
+        var originalPartition = new Geometry("a");
+        originalPartition.AddAnnotations([moved, replaced, nodeWithAnotherId]);
+
+        var clonedPartition = ClonePartition(originalPartition);
+
+        var newIndex = 1;
+        var oldIndex = 0;
+        var notification = new AnnotationMovedAndReplacedInSameParentNotification(newIndex, moved, originalPartition, oldIndex,
+            nodeWithAnotherId, new NumericNotificationId("annotationMovedAndReplaced", 0));
+
+        Assert.ThrowsExactly<InvalidNotificationException>(() =>
+        {
+            CreatePartitionReplicator(clonedPartition, notification);
+        });
     }
 
     #endregion